Move Dynatrace metrics exporter for API v1 to a separate package (#2553)
* Move Dynatrace metrics exporter for API v1 to a separate package

* rename dynatrace exporter

* use an enum for Dynatrace API versions.

* make DynatraceNamingConvention public

* fix formatting

* rename classes to have version at the end, move enum to uppercase, remove thread factory from exporter

* move naming convention and delegate to the v1 implementation

* update config validation and tests

* rename variables in test files

* remove unused imports

* split out validation

* re-add the apiVersion check and test

* update misleading comment and remove duplicate apiVersion check
