+++ implementations/micrometer-registry-dynatrace/src/main/java/io/micrometer/dynatrace/DynatraceNamingConvention.java
@@ -1,5 +1,5 @@
 /**
- * Copyright 2017 VMware, Inc.
+ * Copyright 2021 VMware, Inc.
  * <p>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,30 +18,28 @@ package io.micrometer.dynatrace;
 import io.micrometer.core.instrument.Meter;
 import io.micrometer.core.instrument.config.NamingConvention;
 import io.micrometer.core.lang.Nullable;
-import io.micrometer.core.util.internal.logging.WarnThenDebugLogger;
-
-import java.util.regex.Pattern;
+import io.micrometer.dynatrace.v1.DynatraceNamingConventionV1;
 
 /**
- * {@link NamingConvention} for Dynatrace.
+ * {@link NamingConvention} for Dynatrace. Delegates to the API-specific naming convention.
  *
  * @author Oriol Barcelona Palau
  * @author Jon Schneider
  * @author Johnny Lim
+ * @author Georg Pirklbauer
  * @since 1.1.0
  */
 public class DynatraceNamingConvention implements NamingConvention {
+    private final NamingConvention namingConvention;
 
-    private static final WarnThenDebugLogger logger = new WarnThenDebugLogger(DynatraceNamingConvention.class);
-
-    private static final Pattern NAME_CLEANUP_PATTERN = Pattern.compile("[^\\w._-]");
-    private static final Pattern LEADING_NUMERIC_PATTERN = Pattern.compile("[._-]([\\d])+");
-    private static final Pattern KEY_CLEANUP_PATTERN = Pattern.compile("[^\\w.-]");
-
-    private final NamingConvention delegate;
+    public DynatraceNamingConvention(NamingConvention delegate, DynatraceApiVersion version) {
+        // if (version == DynatraceApiVersion.V1) ...
+        // for now, this check does not make sense, but it will when more naming conventions are added.
+        this.namingConvention = new DynatraceNamingConventionV1(delegate);
+    }
 
     public DynatraceNamingConvention(NamingConvention delegate) {
-        this.delegate = delegate;
+        this(delegate, DynatraceApiVersion.V1);
     }
 
     public DynatraceNamingConvention() {
@@ -50,25 +48,22 @@ public class DynatraceNamingConvention implements NamingConvention {
 
     @Override
     public String name(String name, Meter.Type type, @Nullable String baseUnit) {
-        return "custom:" + sanitizeName(delegate.name(name, type, baseUnit));
+        return namingConvention.name(name, type, baseUnit);
     }
 
-    private String sanitizeName(String name) {
-        if (name.equals("system.load.average.1m")) {
-            return "system.load.average.oneminute";
-        }
-        String sanitized = NAME_CLEANUP_PATTERN.matcher(name).replaceAll("_");
-        if (LEADING_NUMERIC_PATTERN.matcher(sanitized).find()) {
-            logger.log("'" + sanitized + "' (original name: '" + name + "') is not a valid meter name. "
-                    + "Dynatrace doesn't allow leading numeric characters after non-alphabets. "
-                    + "Please rename it to conform to the constraints. "
-                    + "If it comes from a third party, please use MeterFilter to rename it.");
-        }
-        return sanitized;
+    @Override
+    public String name(String name, Meter.Type type) {
+        return namingConvention.name(name, type);
     }
 
     @Override
     public String tagKey(String key) {
-        return KEY_CLEANUP_PATTERN.matcher(delegate.tagKey(key)).replaceAll("_");
+        return namingConvention.tagKey(key);
+    }
+
+    @Override
+    public String tagValue(String value) {
+        return namingConvention.tagValue(value);
     }
 }
+
