+++ implementations/micrometer-registry-dynatrace/src/main/java/io/micrometer/dynatrace/v1/DynatraceCustomMetric.java
@@ -1,0 +1,35 @@
+/**
+ * Copyright 2021 VMware, Inc.
+ * <p>
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * <p>
+ * https://www.apache.org/licenses/LICENSE-2.0
+ * <p>
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.micrometer.dynatrace.v1;
+
+class DynatraceCustomMetric {
+    private final DynatraceMetricDefinition metricDefinition;
+    private final DynatraceTimeSeries timeSeries;
+
+    DynatraceCustomMetric(final DynatraceMetricDefinition metricDefinition, final DynatraceTimeSeries timeSeries) {
+        this.metricDefinition = metricDefinition;
+        this.timeSeries = timeSeries;
+    }
+
+    DynatraceMetricDefinition getMetricDefinition() {
+        return metricDefinition;
+    }
+
+    DynatraceTimeSeries getTimeSeries() {
+        return timeSeries;
+    }
+}
