+++ implementations/micrometer-registry-dynatrace/src/test/java/io/micrometer/dynatrace/v1/DynatraceExporterV1Test.java
@@ -13,12 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micrometer.dynatrace;
+package io.micrometer.dynatrace.v1;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.micrometer.core.instrument.*;
 import io.micrometer.core.instrument.config.validate.ValidationException;
 import io.micrometer.core.ipc.http.HttpSender;
+import io.micrometer.dynatrace.DynatraceApiVersion;
+import io.micrometer.dynatrace.DynatraceConfig;
+import io.micrometer.dynatrace.DynatraceMeterRegistry;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
@@ -40,9 +43,10 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
  *
  * @author Johnny Lim
  */
-class DynatraceMeterRegistryTest {
+class DynatraceExporterV1Test {
 
     private final DynatraceMeterRegistry meterRegistry = createMeterRegistry();
+    private final DynatraceExporterV1 exporter = createExporter();
 
     private final ObjectMapper mapper = new ObjectMapper();
 
@@ -108,14 +112,14 @@ class DynatraceMeterRegistryTest {
 
     @Test
     void putCustomMetricOnSuccessShouldAddMetricIdToCreatedCustomMetrics() throws NoSuchFieldException, IllegalAccessException {
-        Field createdCustomMetricsField = DynatraceMeterRegistry.class.getDeclaredField("createdCustomMetrics");
+        Field createdCustomMetricsField = DynatraceExporterV1.class.getDeclaredField("createdCustomMetrics");
         createdCustomMetricsField.setAccessible(true);
         @SuppressWarnings("unchecked")
-        Set<String> createdCustomMetrics = (Set<String>) createdCustomMetricsField.get(meterRegistry);
+        Set<String> createdCustomMetrics = (Set<String>) createdCustomMetricsField.get(exporter);
         assertThat(createdCustomMetrics).isEmpty();
 
         DynatraceMetricDefinition customMetric = new DynatraceMetricDefinition("metricId", null, null, null, new String[]{"type"}, null);
-        meterRegistry.putCustomMetric(customMetric);
+        exporter.putCustomMetric(customMetric);
         assertThat(createdCustomMetrics).containsExactly("metricId");
     }
 
@@ -123,14 +127,14 @@ class DynatraceMeterRegistryTest {
     void writeMeterWithGauge() {
         meterRegistry.gauge("my.gauge", 1d);
         Gauge gauge = meterRegistry.find("my.gauge").gauge();
-        assertThat(meterRegistry.writeMeter(gauge)).hasSize(1);
+        assertThat(exporter.writeMeter(gauge)).hasSize(1);
     }
 
     @Test
     void writeMeterWithGaugeShouldDropNanValue() {
         meterRegistry.gauge("my.gauge", Double.NaN);
         Gauge gauge = meterRegistry.find("my.gauge").gauge();
-        assertThat(meterRegistry.writeMeter(gauge)).isEmpty();
+        assertThat(exporter.writeMeter(gauge)).isEmpty();
     }
 
     @SuppressWarnings("unchecked")
@@ -139,10 +143,10 @@ class DynatraceMeterRegistryTest {
         AtomicBoolean first = new AtomicBoolean(true);
         meterRegistry.gauge("my.gauge", first, (b) -> b.getAndSet(false) ? 1d : Double.NaN);
         Gauge gauge = meterRegistry.find("my.gauge").gauge();
-        Stream<DynatraceMeterRegistry.DynatraceCustomMetric> stream = meterRegistry.writeMeter(gauge);
-        List<DynatraceMeterRegistry.DynatraceCustomMetric> metrics = stream.collect(Collectors.toList());
+        Stream<DynatraceCustomMetric> stream = exporter.writeMeter(gauge);
+        List<DynatraceCustomMetric> metrics = stream.collect(Collectors.toList());
         assertThat(metrics).hasSize(1);
-        DynatraceMeterRegistry.DynatraceCustomMetric metric = metrics.get(0);
+        DynatraceCustomMetric metric = metrics.get(0);
         DynatraceTimeSeries timeSeries = metric.getTimeSeries();
         try {
             Map<String, Object> map = mapper.readValue(timeSeries.asJson(), Map.class);
@@ -157,11 +161,11 @@ class DynatraceMeterRegistryTest {
     void writeMeterWithGaugeShouldDropInfiniteValues() {
         meterRegistry.gauge("my.gauge", Double.POSITIVE_INFINITY);
         Gauge gauge = meterRegistry.find("my.gauge").gauge();
-        assertThat(meterRegistry.writeMeter(gauge)).isEmpty();
+        assertThat(exporter.writeMeter(gauge)).isEmpty();
 
         meterRegistry.gauge("my.gauge", Double.NEGATIVE_INFINITY);
         gauge = meterRegistry.find("my.gauge").gauge();
-        assertThat(meterRegistry.writeMeter(gauge)).isEmpty();
+        assertThat(exporter.writeMeter(gauge)).isEmpty();
     }
 
     @Test
@@ -169,7 +173,7 @@ class DynatraceMeterRegistryTest {
         AtomicReference<Double> obj = new AtomicReference<>(1d);
         meterRegistry.more().timeGauge("my.timeGauge", Tags.empty(), obj, TimeUnit.SECONDS, AtomicReference::get);
         TimeGauge timeGauge = meterRegistry.find("my.timeGauge").timeGauge();
-        assertThat(meterRegistry.writeMeter(timeGauge)).hasSize(1);
+        assertThat(exporter.writeMeter(timeGauge)).hasSize(1);
     }
 
     @Test
@@ -177,7 +181,7 @@ class DynatraceMeterRegistryTest {
         AtomicReference<Double> obj = new AtomicReference<>(Double.NaN);
         meterRegistry.more().timeGauge("my.timeGauge", Tags.empty(), obj, TimeUnit.SECONDS, AtomicReference::get);
         TimeGauge timeGauge = meterRegistry.find("my.timeGauge").timeGauge();
-        assertThat(meterRegistry.writeMeter(timeGauge)).isEmpty();
+        assertThat(exporter.writeMeter(timeGauge)).isEmpty();
     }
 
     @Test
@@ -185,12 +189,12 @@ class DynatraceMeterRegistryTest {
         AtomicReference<Double> obj = new AtomicReference<>(Double.POSITIVE_INFINITY);
         meterRegistry.more().timeGauge("my.timeGauge", Tags.empty(), obj, TimeUnit.SECONDS, AtomicReference::get);
         TimeGauge timeGauge = meterRegistry.find("my.timeGauge").timeGauge();
-        assertThat(meterRegistry.writeMeter(timeGauge)).isEmpty();
+        assertThat(exporter.writeMeter(timeGauge)).isEmpty();
 
         obj = new AtomicReference<>(Double.NEGATIVE_INFINITY);
         meterRegistry.more().timeGauge("my.timeGauge", Tags.empty(), obj, TimeUnit.SECONDS, AtomicReference::get);
         timeGauge = meterRegistry.find("my.timeGauge").timeGauge();
-        assertThat(meterRegistry.writeMeter(timeGauge)).isEmpty();
+        assertThat(exporter.writeMeter(timeGauge)).isEmpty();
     }
 
     @Test
@@ -198,11 +202,11 @@ class DynatraceMeterRegistryTest {
         Double number = 1d;
         meterRegistry.gauge("my.gauge", number);
         Gauge gauge = meterRegistry.find("my.gauge").gauge();
-        Stream<DynatraceMeterRegistry.DynatraceCustomMetric> series = meterRegistry.writeMeter(gauge);
+        Stream<DynatraceCustomMetric> series = exporter.writeMeter(gauge);
         List<DynatraceTimeSeries> timeSeries = series
-                .map(DynatraceMeterRegistry.DynatraceCustomMetric::getTimeSeries)
+                .map(DynatraceCustomMetric::getTimeSeries)
                 .collect(Collectors.toList());
-        List<DynatraceBatchedPayload> entries = meterRegistry.createPostMessages("my.type", null, timeSeries);
+        List<DynatraceBatchedPayload> entries = exporter.createPostMessages("my.type", null, timeSeries);
         assertThat(entries).hasSize(1);
         assertThat(entries.get(0).metricCount).isEqualTo(1);
         assertThat(isValidJson(entries.get(0).payload)).isEqualTo(true);
@@ -210,14 +214,14 @@ class DynatraceMeterRegistryTest {
 
     @Test
     void whenAllTsTooLargeEmptyMessageListReturned() {
-        List<DynatraceBatchedPayload> messages = meterRegistry.createPostMessages("my.type", null, Collections.singletonList(createTimeSeriesWithDimensions(10_000)));
+        List<DynatraceBatchedPayload> messages = exporter.createPostMessages("my.type", null, Collections.singletonList(createTimeSeriesWithDimensions(10_000)));
         assertThat(messages).isEmpty();
     }
 
     @Test
     void splitsWhenExactlyExceedingMaxByComma() {
         // comma needs to be considered when there is more than one time series
-        List<DynatraceBatchedPayload> messages = meterRegistry.createPostMessages("my.type", "my.group",
+        List<DynatraceBatchedPayload> messages = exporter.createPostMessages("my.type", "my.group",
                 // Max bytes: 15330 (excluding header/footer, 15360 with header/footer)
                 Arrays.asList(createTimeSeriesWithDimensions(750), // 14861 bytes
                         createTimeSeriesWithDimensions(23, "asdfg"), // 469 bytes (overflows due to comma)
@@ -234,7 +238,7 @@ class DynatraceMeterRegistryTest {
 
     @Test
     void countsPreviousAndNextComma() {
-        List<DynatraceBatchedPayload> messages = meterRegistry.createPostMessages("my.type", null,
+        List<DynatraceBatchedPayload> messages = exporter.createPostMessages("my.type", null,
                 // Max bytes: 15330 (excluding header/footer, 15360 with header/footer)
                 Arrays.asList(createTimeSeriesWithDimensions(750), // 14861 bytes
                         createTimeSeriesWithDimensions(10, "asdf"), // 234 bytes + comma
@@ -253,7 +257,7 @@ class DynatraceMeterRegistryTest {
         Measurement measurement3 = new Measurement(() -> Double.NaN, Statistic.VALUE);
         List<Measurement> measurements = Arrays.asList(measurement1, measurement2, measurement3);
         Meter meter = Meter.builder("my.meter", Meter.Type.GAUGE, measurements).register(this.meterRegistry);
-        assertThat(meterRegistry.writeMeter(meter)).isEmpty();
+        assertThat(exporter.writeMeter(meter)).isEmpty();
     }
 
     @Test
@@ -265,7 +269,7 @@ class DynatraceMeterRegistryTest {
         Measurement measurement5 = new Measurement(() -> 2d, Statistic.VALUE);
         List<Measurement> measurements = Arrays.asList(measurement1, measurement2, measurement3, measurement4, measurement5);
         Meter meter = Meter.builder("my.meter", Meter.Type.GAUGE, measurements).register(this.meterRegistry);
-        assertThat(meterRegistry.writeMeter(meter)).hasSize(2);
+        assertThat(exporter.writeMeter(meter)).hasSize(2);
     }
 
     private DynatraceTimeSeries createTimeSeriesWithDimensions(int numberOfDimensions) {
@@ -283,7 +287,22 @@ class DynatraceMeterRegistryTest {
     }
 
     private DynatraceMeterRegistry createMeterRegistry() {
-        DynatraceConfig config = new DynatraceConfig() {
+        DynatraceConfig config = createDynatraceConfig();
+
+        return DynatraceMeterRegistry.builder(config)
+                .httpClient(request -> new HttpSender.Response(200, null))
+                .build();
+    }
+
+    private DynatraceExporterV1 createExporter() {
+        DynatraceConfig config = createDynatraceConfig();
+
+        return new DynatraceExporterV1(config, Clock.SYSTEM, request -> new HttpSender.Response(200, null));
+
+    }
+
+    private DynatraceConfig createDynatraceConfig() {
+        return new DynatraceConfig() {
             @Override
             public String get(String key) {
                 return null;
@@ -303,10 +322,12 @@ class DynatraceMeterRegistryTest {
             public String apiToken() {
                 return "apiToken";
             }
+
+            @Override
+            public DynatraceApiVersion apiVersion() {
+                return DynatraceApiVersion.V1;
+            }
         };
-        return DynatraceMeterRegistry.builder(config)
-                .httpClient(request -> new HttpSender.Response(200, null))
-                .build();
     }
 
     private boolean isValidJson(String json) {
@@ -317,5 +338,4 @@ class DynatraceMeterRegistryTest {
             return false;
         }
     }
-
 }
